refactored: move files to EssentialFeed target
