+++ EssentialFeed/EssentialFeed/Feed Presentation/Comments/FeedImageCommentLoadingViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  FeedImageCommentLoadingViewModel.swift
+//  EssentialFeed
+//
+//  Created by Danil Vassyakin on 3/2/21.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+import Foundation
+
+public struct FeedImageCommentLoadingViewModel {
+	public let isLoading: Bool
+}
