+++ EssentialFeed/EssentialFeed/Feed Presentation/pt-BR.lproj/Comments.strings
